backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0
psycopg2-binary>=2.9,<3.0
python-decouple>=3.8
Pillow>=10.0
gunicorn>=22.0
whitenoise>=6.7
djangorestframework-simplejwt>=5.3
dj-database-url>=2.2
